requirements.txt
django==2.2
django-environ==0.4.1
djangorestframework==3.10.3
python-decouple==3.3
dj-database-url==0.4.2
coverage==5.2.1
